update 4 files Schinese  (#36)
* update Chinese translation

I retranslated the "bg2-autumns-twilight-mod-1_3_0" file, and thoroughly translated and proofread all the text. At the same time, the (Autumns_Twilight.tp2) file was modified. The (Chinese(Simplified)) folder in the MOD was renamed to (SChinese).

* Update P_DuskJ.tra

Updated the SChinese translation of a sentence.

* update SCHINESE translation

update SCHINESE translation

* update setup.tra schinese

update setup.tra schinese

* update 3 files schinese

 update 3 files schinese